2024_AI_DEV_SUMMIT/MachineLearningWithFastai/answer-questions-in-meetings/requirements.txt
chardet==5.2.0
packaging==24.0
fastai==2.7.15
protobuf==5.26.1
nltk==3.8.1
datasets==2.19.1
pandas==2.2.2
transformers==4.40.2
accelerate==0.30.1
beautifulsoup4==4.12.3
sentencepiece==0.2.0

# deepgram for transcription
deepgram-sdk==3.2.7
verboselogs
python-dotenv
pyaudio

# openai
openai==1.30.1

# for playing audio
playsound==1.3.0
PyObjC==10.2
